*Removed server.js merged the stuff back into app.js as is the express way
